refactor(domain): unificar albumId a Long en todas las capas (M6)
El id de Album ya era Long punta a punta salvo en las asociaciones
album->musico/colector, tipadas como Int. Se unifica a Long en
repositorios/Impl/use cases/VMs/UiState/@Path/fakes/tests y se eliminan
los .toInt()/.toLong() ahora redundantes sobre ids de Album. Sin cambio
de comportamiento (solo widening). Value class descartado por
intratabilidad transversal (Room/Retrofit/Gson/SSH fixtures).

+++ app/src/androidTest/java/com/misw4203/vinilos/di/FakeCollectorRepository.kt
@@ -16,7 +16,7 @@ class FakeCollectorRepository @Inject constructor() : CollectorRepository {
         CollectorSummary(2, "María Alejandra Palacios", "3502889087", "j.palacios@outlook.es"),
     )
 
-    override suspend fun addAlbumToCollector(collectorId: Int, albumId: Int, price: Double, status: String) = Unit
+    override suspend fun addAlbumToCollector(collectorId: Int, albumId: Long, price: Double, status: String) = Unit
 
     override suspend fun getCollectorDetail(id: Int): CollectorDetail = CollectorDetail(
         id = id,
@@ -56,5 +56,5 @@ class FakeCollectorRepository @Inject constructor() : CollectorRepository {
 
     override suspend fun removeFavoriteMusician(collectorId: Int, musicianId: Int) {}
     override suspend fun removeFavoriteBand(collectorId: Int, bandId: Int) {}
-    override suspend fun removeAlbumFromCollector(collectorId: Int, albumId: Int) {}
+    override suspend fun removeAlbumFromCollector(collectorId: Int, albumId: Long) {}
 }
