+++ app/src/main/java/com/misw4203/vinilos/presentation/viewmodel/AddAlbumToBandViewModel.kt
@@ -55,14 +55,14 @@ class AddAlbumToBandViewModel @Inject constructor(
         )
     }
 
-    fun onAddAlbum(albumId: Int) {
+    fun onAddAlbum(albumId: Long) {
         if (_uiState.value is AddAlbumToBandUiState.Adding) return
         _uiState.value = AddAlbumToBandUiState.Adding(albumId)
         viewModelScope.launch {
-            when (runCatchingDomain { addAlbumToBand(bandId, albumId.toLong()) }) {
+            when (runCatchingDomain { addAlbumToBand(bandId, albumId) }) {
                 is DomainResult.Ok -> {
-                    val album = _form.value.allAlbums.firstOrNull { it.id == albumId.toLong() }
-                    val newIds = _form.value.currentAlbumIds + albumId.toLong()
+                    val album = _form.value.allAlbums.firstOrNull { it.id == albumId }
+                    val newIds = _form.value.currentAlbumIds + albumId
                     _form.value = _form.value.copy(
                         currentAlbumIds = newIds,
                         filteredAvailable = computeFiltered(_form.value.allAlbums, newIds, _form.value.query),
