+++ app/src/test/java/com/misw4203/vinilos/presentation/viewmodel/AddAlbumToMusicianViewModelTest.kt
@@ -38,7 +38,7 @@ class AddAlbumToMusicianViewModelTest {
     ) : MusicianRepository {
         override suspend fun getMusicians(): List<MusicianSummary> = emptyList()
         override suspend fun getMusicianDetail(id: Int): Musician = musicianResult.getOrThrow()
-        override suspend fun addAlbumToMusician(musicianId: Int, albumId: Int) = addResult.getOrThrow()
+        override suspend fun addAlbumToMusician(musicianId: Int, albumId: Long) = addResult.getOrThrow()
         override suspend fun addPrizeToMusician(musicianId: Int, prizeId: Int, premiationDate: String) = Unit
     }
 
@@ -121,9 +121,9 @@ class AddAlbumToMusicianViewModelTest {
             advanceUntilIdle()
             assertEquals(AddAlbumToMusicianUiState.Ready, awaitItem())
 
-            vm.onAddAlbum(2)
+            vm.onAddAlbum(2L)
 
-            assertEquals(AddAlbumToMusicianUiState.Adding(2), awaitItem())
+            assertEquals(AddAlbumToMusicianUiState.Adding(2L), awaitItem())
             advanceUntilIdle()
             assertEquals(AddAlbumToMusicianUiState.Ready, awaitItem())
             cancelAndIgnoreRemainingEvents()
@@ -144,9 +144,9 @@ class AddAlbumToMusicianViewModelTest {
             assertEquals(AddAlbumToMusicianUiState.Ready, awaitItem())
 
             vm.events.test {
-                vm.onAddAlbum(2)
+                vm.onAddAlbum(2L)
                 // Adding is set synchronously on uiState, not emitted as an event.
-                assertEquals(AddAlbumToMusicianUiState.Adding(2), vm.uiState.value)
+                assertEquals(AddAlbumToMusicianUiState.Adding(2L), vm.uiState.value)
                 advanceUntilIdle()
                 capturedEvent = awaitItem()
                 cancelAndIgnoreRemainingEvents()
