+++ app/src/test/java/com/misw4203/vinilos/presentation/viewmodel/AddMusiciansToBandViewModelTest.kt
@@ -37,7 +37,7 @@ class AddMusiciansToBandViewModelTest {
         var allMusicians: List<MusicianSummary> = emptyList()
         override suspend fun getMusicians(): List<MusicianSummary> = allMusicians
         override suspend fun getMusicianDetail(id: Int): Musician = error("not used")
-        override suspend fun addAlbumToMusician(musicianId: Int, albumId: Int) = Unit
+        override suspend fun addAlbumToMusician(musicianId: Int, albumId: Long) = Unit
         override suspend fun addPrizeToMusician(musicianId: Int, prizeId: Int, premiationDate: String) = Unit
     }
